test: split kernel generator parity checks into separate file
